Changed methods for validation.

+++ validators/plane_validator.py
@@ -5,7 +5,7 @@ from .base_validator import ShapeBase
 
 class PlaneValidator(ShapeBase):
 
-    width: int = Field(gt=0, default=2)
-    depth: int = Field(gt=0, default=2)
+    width: float = Field(gt=0, default=2)
+    depth: float = Field(gt=0, default=2)
     segs_w: int = Field(gt=1, default=6)
     segs_d: int = Field(gt=1, default=6)
